# Rework DG/HDG execution logic to limit warehouse and shoudlExecute calls